requirements.txt
# 项目依赖包
# 安装方式: pip install -r requirements.txt

# 测试框架
pytest>=7.4.0
pytest-html>=3.2.0

# HTTP请求
requests>=2.31.0

# 配置文件解析
PyYAML>=6.0

# 环境变量
python-dotenv>=1.0.0

# Mock服务
Flask>=2.3.0

# 类型提示（可选）
typing-extensions>=4.8.0

